refactor(routes): refactor blog routes into academy + delete unused files
