+++ deriv.com.conf
@@ -29,6 +29,10 @@ server {
         return 301  https://$http_host/careers/;
     }
 
+    location ~* "^/([\w]{2}|zh-tw|zh-cn)/academy" {
+        return 301  https://$http_host/academy/;
+    }
+
     location = /sw.js {
         add_header Cache-Control "public, max-age=0, must-revalidate";
     }
