+++ src/components/custom/other-platforms.js
@@ -527,9 +527,9 @@ export const NavResources = ({ onClick }) => (
         />
         <CardLink
             icon={() => <img src={Blog} alt="" width="24" height="24" />}
-            title={localize('Blog')}
+            title={localize('Academy')}
             onClick={onClick}
-            to="/blog"
+            to="/academy/"
         />
     </Flex>
 )
