+++ src/components/elements/off-canvas-menu.js
@@ -458,11 +458,11 @@ export const OffCanvasMenuWrapper = (props) => {
                                 <img src={Diagonal} alt="" width="16" height="16" />
                             </SpanSvg>
                         </StyledLink>
-                        <StyledLink to="/blog" onClick={handleArrowClick}>
+                        <StyledLink to="/academy/" onClick={handleArrowClick}>
                             <div>
                                 <img src={Blog} alt="" width="24" height="24" />
                             </div>
-                            <Span>{localize('Blog')}</Span>
+                            <Span>{localize('Academy')}</Span>
                             <SpanSvg>
                                 <img src={Diagonal} alt="" width="16" height="16" />
                             </SpanSvg>
