+++ src/components/layout/footer/main-links.js
@@ -186,7 +186,7 @@ const MainLinksSection = ({ is_ppc, is_ppc_redirect, is_eu_country }) => {
                             </Link>
                         </LinkWrapper>
                         <LinkWrapper>
-                            <Link to="/blog">{localize('Blog')}</Link>
+                            <Link to="/academy/">{localize('Academy')}</Link>
                         </LinkWrapper>
                     </LinksCol>
                 </Flex>
