+++ src/pages/academy/common/_styles.js
