+++ src/pages/academy/components/_banner.js
