+++ src/pages/academy/components/_common.js
