+++ src/pages/academy/components/_hero.js
