+++ src/pages/academy/components/_video-player.js
