+++ src/pages/academy/components/recent-featured-posts/_style.js
