+++ src/pages/academy/components/video-banner/carousel/_carousel-style.js
