+++ src/pages/academy/videos/_video-card.js
