--- src/pages/blog/_markets_news_data.js
@@ -1,68 +1,0 @@
-export const market_news_data = [
-    {
-        report_type: "Weekly report",
-        date: "24 May 2021 - 30 May 2021",
-        read_time: 3,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Weekly report",
-        date: "17 May 2021 - 24 May 2021",
-        read_time: 4,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Daily report",
-        date: "17 May 2021 - 24 May 2021",
-        read_time: 4,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Monthly report",
-        date: "17 May 2021 - 24 May 2021",
-        read_time: 4,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Weekly report",
-        date: "10 May 2021 - 17 May 2021",
-        read_time: 3,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Yearly report",
-        date: "10 May 2021 - 17 May 2021",
-        read_time: 3,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Weekly report",
-        date: "5 April 2021 - 12 April 2021",
-        read_time: 2,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Weekly report",
-        date: "22 March 2021 - 29 March 2021",
-        read_time: 6,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Weekly report",
-        date: "22 March 2021 - 29 March 2021",
-        read_time: 2,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Weekly report",
-        date: "1 March 2021 - 8 March 2021",
-        read_time: 3,
-        img_source: "https://source.unsplash.com/random/104x78"
-    },
-    {
-        report_type: "Weekly report",
-        date: "1 March 2021 - 8 March 2021",
-        read_time: 4,
-        img_source: "https://source.unsplash.com/random/104x78"
-    }
-]
