--- src/pages/blog/template/_side-cta/_data.js
@@ -1,7 +1,0 @@
-import Stocks from './_images/stocks.png'
-
-export const pdf_data = {
-    image: Stocks,
-    title: 'How to trade stocks - the smart way',
-    author: 'Vince Stanzione',
-}
